requirements.txt
alembic==1.13.0
SQLAlchemy==2.0.23
python-decouple==3.8
psycopg2-binary==2.9.9